requirements.txt
python-dotenv
openai
sentense_transformers
langchain
qdrant_client
Flask
waitress